Refactor database connection and file structure
